+++ DRSRule.psm1
@@ -1,28 +1,90 @@
-#.ExternalHelp DRSRule.Help.xml
-Function Get-DrsVMGroup {
-  [CmdletBinding()]
+<#  .Description
+    Retrieves the DRS VM groups. It returns DRS VM groups that correspond to the filter criteria provided by the cmdlet parameters.
+
+    The default return type holds more information than the "raw" DRS object that vSphere uses.  There is also a switch to allow for just returning said raw DRS object, quite useful for consumption by other cmdlets in this module.
+
+    .Synopsis
+    This cmdlet retrieves the DRS VM groups
+
+    .Example
+    Get-DrsVMGroup -Name '*VM Group 1*'
+    Name               Cluster         UserCreated        VM
+    ----               -------         -----------        --
+    VM Group 1         Cluster1        True               {VM1,VM2}
+    VM Group 12        Cluster1        True               {VM3}
+    New VM Group 100   Cluster2        True               {VM4,VM5,VM6}
+
+    Returns all DRS VM groups with name like '*VM Group 1*'
+
+    .Example
+    Get-DrsVMGroup -Cluster Cluster1 -Name 'VM Group 1'
+    Name               Cluster         UserCreated        VM
+    ----               -------         -----------        --
+    VM Group 1         Cluster1        True               {VM1,VM2}
+
+    The DRS VM group with the exact name 'VM Group 1' from Cluster1 will be returned
+
+    .Example
+    Get-Cluster Cluster2 | Get-DrsVMGroup
+    Name               Cluster         UserCreated        VM
+    ----               -------         -----------        --
+    VM Group 5         Cluster2        True               {VM101,VM102}
+    testVMGroup        Cluster2        True               {VM0,VM1001,VM1002}
+
+    Returns all DRS VM groups in cluster "Cluster2"
+
+    .Example
+    Get-VM DrsRuleTest1 | Get-DrsVMGroup
+    Name             Cluster   UserCreated   VM
+    ----             -------   -----------   --
+    TestVMGroup1     myClus0   True          {DrsRuleTest1, DrsRuleTest0}
+
+    Gets a DRS VMGroup by the related VM object. Returns the VMGroup(s) of which a VM is a part, if any
+
+    .Outputs
+    If corrsponding DRS VMGroup(s) found, either DRSRule.VMGroup in "normal" mode, or VMware.Vim.ClusterVmGroup in "-ReturnRaw" mode. Else, $null
+
+    .Link
+    https://github.com/PowerCLIGoodies/DRSRule
+    New-DrsVMGroup
+    Remove-DrsVMGroup
+    Set-DrsVMGroup
+#>
+function Get-DrsVMGroup {
+  [CmdletBinding(DefaultParameterSetName = "ByName")]
   [OutputType([DRSRule.VMGroup],[VMware.Vim.ClusterVmGroup])]
   param(
-    [Parameter(Position = 0)]
+    ## Name of DRS VM Group to get (or, all if no name specified)
+    [Parameter(Position = 0, ParameterSetName="ByName")]
     [string]${Name} = '*',
 
-    [Parameter(Position = 1, ValueFromPipeline = $True)]
+    ## Cluster from which to get DRS VM group (or, all clusters if no name specified)
+    [Parameter(Position = 1, ParameterSetName="ByName", ValueFromPipeline=$True)]
     [PSObject[]]${Cluster},
 
+    ## Virtual Machine for which to get the corresponding DRS VMGroup(s), if any
+    [Parameter(Position = 0, Mandatory=$true, ValueFromPipeline=$true, ParameterSetName="ByRelatedObject")]
+    [VMware.VimAutomation.Types.VirtualMachine]$VM,
+
+    ## Switch:  return "raw" VMware.Vim.ClusterVmGroup object (contains less info, but useful to other functions that can consume this raw object)
     [switch]$ReturnRawGroup
   )
 
-  Process{
+  Process {
+    ## is this invocation getting item by related object?
+    $bByRelatedObject = $PSCmdlet.ParameterSetName -eq "ByRelatedObject"
     ## get cluster object(s) from the Cluster param (if no value was specified -- gets all clusters)
-    Get-ClusterObjFromClusterParam -Cluster $Cluster | ForEach-Object -Process {
+    $arrClustersToCheck = if ($bByRelatedObject) {$VM.VMHost.Parent} else {Get-ClusterObjFromClusterParam -Cluster $Cluster}
+    ## for the cluster(s) to check, try to get the pertinent VMGroups
+    $arrClustersToCheck | ForEach-Object -Process {
       $oThisCluster = $_
       ## update the View data, in case it was stale
       $oThisCluster.ExtensionData.UpdateViewData("ConfigurationEx")
       ## foreach ClusterVmGroup item, return something
       $oThisCluster.ExtensionData.ConfigurationEx.Group |
       Where-Object -FilterScript {
-        ## changed to "like" from "match" -- "*" with -match causes error, as it is expecting regex, not just std wildcard
-        ($_ -is [VMware.Vim.ClusterVmGroup]) -and ($_.Name -like ${Name})
+        ## where it's the given type, and, if ByRelatedObject, it contains the MoRef of this realted object -- else, if the name is like the specified name
+        ($_ -is [VMware.Vim.ClusterVmGroup]) -and $(if ($bByRelatedObject) {$_.VM -contains $VM.Id} else {$_.Name -like ${Name}})
       } |
       ForEach-Object -Process {
         if ($true -eq $ReturnRawGroup) {return $_}
@@ -45,31 +107,97 @@ Function Get-DrsVMGroup {
   }
 }
 
-#.ExternalHelp DRSRule.Help.xml
-Function Get-DrsVMHostGroup {
-  [CmdletBinding()]
+
+<#  .Description
+    This cmdlet retrieves the DRS VM groups.
+    It returns DRS VM groups that correspond to the filter criteria provided by the cmdlet parameters.
+
+    The default return type holds more information than the "raw" DRS object that vSphere uses.  There is also a switch to allow for just returning said raw DRS object, quite useful for consumption by other cmdlets in this module.
+
+    .Synopsis
+    Retrieves the DRS VMHost groups
+
+    .Example
+    Get-DRSVMHostGroup -Name '*VMHost Group 1*'
+    Name               Cluster         UserCreated        VMHost
+    ----               -------         -----------        ------
+    VMHost Group 1     Cluster1        True               {esx1,esx2}
+    VMHost Group 12    Cluster1        True               {esx3}
+    New VMHost Group 1 Cluster2        True               {esx4,esx5,esx6}
+
+    Returns all DRS VMHost groups with name like '*VMHost Group 1*'
+
+    .Example
+    Get-DRSVMHostGroup -Cluster Cluster1 -Name 'VMHost Group 1'
+    Name               Cluster         UserCreated        VMHost
+    ----               -------         -----------        ------
+    VMHost Group 1     Cluster1        True               {esx1,esx2}
+
+    The DRS VMHost group with the exact name 'VMHost Group 1' from Cluster1 will be returned
+
+    .Example
+    Get-Cluster Cluster2 | Get-DrsVMHostGroup
+    Name               Cluster         UserCreated        VMHost
+    ----               -------         -----------        ------
+    oldVMHostGroup     Cluster2        True               {esx11,esx12}
+    VMHost Group DR    Cluster2        True               {esx13}
+    New VMHost Grp 3   Cluster2        True               {esx14,esx15,esx16}
+
+    Returns all DRS VMHost groups in cluster "Cluster2"
+
+    .Example
+    Get-VMHost esx11 | Get-DrsVMHostGroup
+    Name             Cluster   UserCreated   VMHost
+    ----             -------   -----------   ------
+    oldVMHostGroup   Cluster2  True          {esx11,esx12}
+
+    Gets a DRS VMHostGroup by the related VMHost object. Returns the VMHostGroup(s) of which a VMHost is a part, if any
+
+    .Outputs
+    If corrsponding DRS VMHostGroup(s) found, either DRSRule.VMHostGroup in "normal" mode, or VMware.Vim.ClusterHostGroup in "-ReturnRaw" mode. Else, $null
+
+    .Link
+    https://github.com/PowerCLIGoodies/DRSRule
+    New-DrsVMHostGroup
+    Remove-DrsVMHostGroup
+    Set-DrsVMHostGroup
+#>
+function Get-DrsVMHostGroup {
+  [CmdletBinding(DefaultParameterSetName = "ByName")]
   [OutputType([DRSRule.VMHostGroup],[VMware.Vim.ClusterHostGroup])]
   param(
-    [Parameter(Position = 0)]
+    ## Name of DRS VMHost Group to get (or, all if no name specified)
+    [Parameter(Position = 0, ParameterSetName="ByName")]
     [ValidateNotNullOrEmpty()]
     [string]${Name} = '*',
 
-    [Parameter(Position = 1, ValueFromPipeline = $True)]
+    ## Cluster from which to get DRS VMHost group (or, all clusters if no name specified)
+    [Parameter(Position = 1, ParameterSetName="ByName", ValueFromPipeline = $True)]
     [ValidateNotNullOrEmpty()]
     [PSObject[]]${Cluster},
 
+    ## VMHost for which to get the corresponding DRS VMHostGroup(s), if any
+    [Parameter(Position = 0, Mandatory=$true, ValueFromPipeline=$true, ParameterSetName="ByRelatedObject")]
+    [VMware.VimAutomation.Types.VMHost]$VMHost,
+
+    ## Switch:  return "raw" VMware.Vim.ClusterHostGroup object (contains less info, but useful to other functions that can consume this raw object)
     [switch]$ReturnRawGroup
   )
 
   Process{
-    Get-ClusterObjFromClusterParam -Cluster $Cluster | ForEach-Object -Process {
+    ## is this invocation getting item by related object?
+    $bByRelatedObject = $PSCmdlet.ParameterSetName -eq "ByRelatedObject"
+    ## get cluster object(s) from the Cluster param (if no value was specified -- gets all clusters)
+    $arrClustersToCheck = if ($bByRelatedObject) {$VMHost.Parent} else {Get-ClusterObjFromClusterParam -Cluster $Cluster}
+    ## for the cluster(s) to check, try to get the pertinent VMHostGroups
+    $arrClustersToCheck | ForEach-Object -Process {
       $oThisCluster = $_
       ## update the View data, in case it was stale
       $oThisCluster.ExtensionData.UpdateViewData("ConfigurationEx")
       ## foreach ClusterVmGroup item, return something
       $oThisCluster.ExtensionData.ConfigurationEx.Group |
       Where-Object -FilterScript {
-        ($_ -is [VMware.Vim.ClusterHostGroup]) -and ($_.Name -like ${Name})
+        ($_ -is [VMware.Vim.ClusterHostGroup]) -and $(if ($bByRelatedObject) {$_.Host -contains $VMHost.Id} else {$_.Name -like ${Name}})
       } |
       ForEach-Object -Process {
         if ($true -eq $ReturnRawGroup) {return $_}
@@ -91,30 +219,88 @@ Function Get-DrsVMHostGroup {
   }
 }
 
-#.ExternalHelp DRSRule.Help.xml
-Function Get-DrsVMToVMRule {
-  [CmdletBinding()]
+
+<#  .Description
+    This cmdlet retrieves the DRS VM to VM rules.
+    It returns DRS VM to VM rules that correspond to the filter criteria provided by the cmdlet parameters.
+    The VM to VM rules can be either affinity- or anti-affinity rules.
+
+    The default return type holds more information than the "raw" DRS object that vSphere uses.  There is also a switch to allow for just returning said raw DRS object, quite useful for consumption by other cmdlets in this module.
+
+    .Synopsis
+    Retrieves the DRS VM to VM rules
+
+    .Example
+    Get-DrsVMToVMRule -Name 'Rule 1*'
+    Name         Cluster         Enabled     KeepTogether  Mandatory   VM
+    ----         -------         -------     ------------  ---------   --
+    Rule 1       Cluster1        False       False         False       {VM1, VM2}
+    Rule 12      Cluster2        True        True          True        {VM3, VM4}
+
+    Returns all DRS VM to VM rules with name like 'Rule 1*'
+
+    .Example
+    Get-Cluster Cluster2 | Get-DrsVMtoVMRule
+    Name         Cluster         Enabled     KeepTogether  Mandatory   VM
+    ----         -------         -------     ------------  ---------   --
+    Rule 0       Cluster2        False       False         False       {VM101, VM102}
+    Rule 11      Cluster2        True        True          True        {VM103, VM014}
+    Rule_old     Cluster2        False       True          True        {VM110, VM111}
+
+    Returns all DRS VM to VM rules in Cluster2
+
+    .Example
+    Get-VM VM3 | Get-DrsVMToVMRule
+    Name         Cluster         Enabled     KeepTogether  Mandatory   VM
+    ----         -------         -------     ------------  ---------   --
+    Rule 12      Cluster2        True        True          True        {VM3, VM4}
+
+    Returns all DRS VM to VM rules involving the specified VM, "VM3"
+
+    .Outputs
+    DRSRule.VMToVMRule object with information about the given DRS VM to VM rule, or a raw vSphere object of on of the types VMware.Vim.ClusterAffinityRuleSpec or VMware.Vim.ClusterAntiAffinityRuleSpec, depending on if the rule is affinity or anti-affinity
+
+    .Link
+    https://github.com/PowerCLIGoodies/DRSRule
+    New-DrsVMToVMRule
+    Remove-DrsVMToVMRule
+    Set-DrsVMToVMRule
+#>
+function Get-DrsVMToVMRule {
+  [CmdletBinding(DefaultParameterSetName = "ByName")]
   [OutputType([DRSRule.VMToVMRule],[VMware.Vim.ClusterAffinityRuleSpec],[VMware.Vim.ClusterAntiAffinityRuleSpec])]
   param(
-    [Parameter(Position = 0)]
+    ## Name of DRS VM-to-VMHost rule to get (or, all if no name specified)
+    [Parameter(Position = 0, ParameterSetName="ByName")]
     [string]${Name} = '*',
 
-    [Parameter(Position = 1, ValueFromPipeline = $True)]
+    ## Cluster from which to get DRS VM-to-VM rule (or, all clusters if no name specified)
+    [Parameter(Position = 1, ParameterSetName="ByName", ValueFromPipeline = $True)]
     [PSObject[]]${Cluster},
 
+    ## Virtual Machine for which to get the corresponding VM-to-VM DRS rule(s), if any
+    [Parameter(Position = 0, Mandatory=$true, ValueFromPipeline=$true, ParameterSetName="ByRelatedObject")]
+    [VMware.VimAutomation.Types.VirtualMachine]$VM,
+
+    ## Switch:  return DRS VM to VM rule as "raw" VMware.Vim.ClusterAffinityRuleSpec or VMware.Vim.ClusterAntiAffinityRuleSpec object (contains less info, but useful to other functions that can consume this raw object)
     [switch]$ReturnRawRule
   )
 
   Process {
-    Get-ClusterObjFromClusterParam -Cluster $Cluster | ForEach-Object -Process {
+    ## is this invocation getting item by related object?
+    $bByRelatedObject = $PSCmdlet.ParameterSetName -eq "ByRelatedObject"
+    ## get cluster object(s) from the Cluster param (if no value was specified -- gets all clusters)
+    $arrClustersToCheck = if ($bByRelatedObject) {$VM.VMHost.Parent} else {Get-ClusterObjFromClusterParam -Cluster $Cluster}
+    ## for the cluster(s) to check, try to get the pertinent VM-to-VM rules
+    $arrClustersToCheck | ForEach-Object -Process {
       $oThisCluster = $_
       ## update the View data, in case it was stale
       $oThisCluster.ExtensionData.UpdateViewData("ConfigurationEx")
       ## foreach rule item, return something
       $oThisCluster.ExtensionData.ConfigurationEx.Rule |
       Where-Object -FilterScript {
-        ($_.Name -like ${Name}) -and
-        ($_ -is [VMware.Vim.ClusterAffinityRuleSpec] -or $_ -is [VMware.Vim.ClusterAntiAffinityRuleSpec])
+        ($_ -is [VMware.Vim.ClusterAffinityRuleSpec] -or $_ -is [VMware.Vim.ClusterAntiAffinityRuleSpec]) -and
+        $(if ($bByRelatedObject) {$_.VM -contains $VM.Id} else {$_.Name -like ${Name}})
       } |
       ForEach-Object -Process {
         if ($ReturnRawRule) {$_}
@@ -140,32 +326,114 @@ Function Get-DrsVMToVMRule {
   }
 }
 
-#.ExternalHelp DRSRule.Help.xml
-Function Get-DrsVMToVMHostRule {
-  [CmdletBinding()]
+
+<#  .Description
+    This cmdlet retrieves the DRS VM to VMHost rules.
+    It returns a number of DRS VM to VMHost rules that correspond to the filter criteria provided by the cmdlet parameters.
+
+    The default return type holds more information than the "raw" DRS object that vSphere uses.  There is also a switch to allow for just returning said raw DRS object, quite useful for consumption by other cmdlets in this module.
+
+    .Synopsis
+    Retrieve the DRS VM to VMHost rules
+
+    .Example
+    Get-DrsVMToVMHostRule -Name 'Rule 1*'
+    Name         Cluster         Enabled     Mandatory   VMGroupName
+    ----         -------         -------     ---------   -----------
+    Rule 1       Cluster1        False       False       VM Group 1
+    Rule 11      Cluster2        True        True        All VM
+
+    Returns all DRS VM to VMHost rules with name like 'Rule 1*'
+
+    .Example
+    Get-DrsVMtoVMHostRule -Cluster Cluster[12]
+    Name         Cluster         Enabled     Mandatory   VMGroupName
+    ----         -------         -------     ---------   -----------
+    Rule 0       Cluster1        True        False       VM Group 1
+    Rule 1       Cluster1        False       False       VM Group 12
+    Rule 2       Cluster1        False       False       VM Group 31
+    Rule 11      Cluster2        True        True        All VM
+    Rule_bak     Cluster2        True        False       testVMGroup
+    Rule_toDel   Cluster2        False       True        VM Group 5
+
+    Returns all DRS VM to VMHost rules in clusters named "Cluster1" and "Cluster2"
+
+    .Example
+    Get-VM myVM0 | Get-DrsVMtoVMHostRule
+    Name         Cluster         Enabled     Mandatory   VMGroupName
+    ----         -------         -------     ---------   -----------
+    Rule 2       Cluster1        False       False       VM Group 31
+
+    Returns all DRS VM to VMHost rules in the cluster in which "myVM0" resides that involve a DRS VMGroup of which "myVM0" is a member
+
+    .Example
+    Get-VMHost myhost0.dom.com | Get-DrsVMtoVMHostRule
+    Name         Cluster         Enabled     Mandatory   VMGroupName
+    ----         -------         -------     ---------   -----------
+    Rule_toDel   Cluster2        False       True        VM Group 5
+
+    Returns all DRS VM to VMHost rules in the cluster of which VMHost "myhost0.dom.com" is a part, and that involve a DRS VMHostGroup of which "myhost0.dom.com" is a member (either as the Affine or AntiAffine VMHost group)
+
+    .Outputs
+    DRSRule.VMToVMHostRule bject with information about the given DRS VM to VMHost rule, or a "raw" VMware.Vim.ClusterVmHostRuleInfo vSphere object
+
+    .Link
+    https://github.com/PowerCLIGoodies/DRSRule
+    New-DrsVMToVMHostRule
+    Remove-DrsVMToVMHostRule
+    Set-DrsVMToVMHostRule
+#>
+function Get-DrsVMToVMHostRule {
+  [CmdletBinding(DefaultParameterSetName = "ByName")]
   [OutputType([DRSRule.VMToVMHostRule],[VMware.Vim.ClusterVmHostRuleInfo])]
   param(
-    [Parameter(Position = 0)]
+    ## Name of DRS VM affinity/antiaffinity rule to get (or, all if no name specified)
+    [Parameter(Position = 0, ParameterSetName="ByName")]
     [ValidateNotNullOrEmpty()]
     [string]${Name} = '*',
 
-    [Parameter(Position = 1, ValueFromPipeline = $True)]
+    ## Cluster from which to get DRS VM-to-VMhost rule (or, all clusters if no name specified)
+    [Parameter(Position = 1, ParameterSetName="ByName", ValueFromPipeline = $True)]
     [PSObject[]]${Cluster},
 
+    ## Virtual Machine for which to get the corresponding VM-to-VMHost DRS rule(s), if any
+    [Parameter(Position = 0, Mandatory=$true, ValueFromPipeline=$true, ParameterSetName="ByRelatedVM")]
+    [VMware.VimAutomation.Types.VirtualMachine]$VM,
+
+    ## VMHost for which to get the corresponding VM-to-VMHost DRS rule(s), if any
+    [Parameter(Position = 0, Mandatory=$true, ValueFromPipeline=$true, ParameterSetName="ByRelatedVMHost")]
+    [VMware.VimAutomation.Types.VMHost]$VMHost,
+
+    ## Switch:  return DRS VM to VMHost rule as "raw" VMware.Vim.ClusterVmHostRuleInfo object (contains less info, but useful to other functions that can consume this raw object)
     [switch]$ReturnRawRule
   )
 
   Process {
-    Get-ClusterObjFromClusterParam -Cluster $Cluster | ForEach-Object -Process {
+    ## is this invocation getting item by related object?
+    $bByRelatedObject = "ByRelatedVM", "ByRelatedVMHost" -contains $PSCmdlet.ParameterSetName
+    ## get cluster object(s) from the Cluster param (if no value was specified -- gets all clusters), and the FilterScript scriptblock to use for the Where-Object call later, for "filtering" rules
+    $arrClustersToCheck, $sbFilterScript = if ($bByRelatedObject) {
+      if ($PSCmdlet.ParameterSetName -eq "ByRelatedVM") {
+        ## the cluster in which the VM resides, and, a scriptblock that checks if the list of names of DRS VMGroups of which this VM is a part contains the VMGroupName property's value in the given VMToVMHost rule
+        $VM.VMHost.Parent, {($VM | Get-DrsVMGroup -ReturnRaw).Name -contains $_.VmGroupName}
+      } else {
+        $arrNamesOfVMHostGroups_thisVMHost = ($VMHost | Get-DrsVMHostGroup -ReturnRaw).Name
+        ## the cluster in which the VMHost resides, and, a scriptblock that checks if the list of names of DRS VMHostGroups of which this VMHost is a part contains either the AffineHostGroupName or AffineHostGroupName property's value in the given VMToVMHost rule
+        #   good way to do it using Compare-Object, but possibly more confusing to read -- checks to see if there is more than zero "equal" items in the two arrays
+        # $VMHost.Parent, {(Compare-Object -ReferenceObject $arrNamesOfVMHostGroups_thisVMHost -DifferenceObject @($_.AffineHostGroupName, $_.AntiAffineHostGroupName) -IncludeEqual -ExcludeDifferent | Measure-Object).Count -gt 0}
+        $VMHost.Parent, {$arrNamesOfVMHostGroups_thisVMHost -contains $_.AffineHostGroupName -or ($arrNamesOfVMHostGroups_thisVMHost -contains $_.AntiAffineHostGroupName)}
+      } ## end else
+    } else {(Get-ClusterObjFromClusterParam -Cluster $Cluster), {$_.Name -like $Name}}
+    ## for the cluster(s) to check, try to get the pertinent VM-to-VMHost rules
+    $arrClustersToCheck | ForEach-Object -Process {
       $oThisCluster = $_
       ## update the View data, in case it was stale
       $oThisCluster.ExtensionData.UpdateViewData("ConfigurationEx")
       ## foreach rule item, return something
       $oThisCluster.ExtensionData.ConfigurationEx.Rule |
-      Where-Object -FilterScript {
-        $_.Name -like ${Name} -and
-        $_ -is [VMware.Vim.ClusterVmHostRuleInfo]
-      } |
+      Where-Object -FilterScript {$_ -is [VMware.Vim.ClusterVmHostRuleInfo]} |
+      ## filter by the VmGroupName, the VMHostGroupName, or by the rule name, depending on the parameters supplied to this cmdlet call
+      Where-Object -FilterScript $sbFilterScript |
       ForEach-Object -Process {
         if ($ReturnRawRule) {$_}
         else {